+++ packages/terra-application-navigation/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 1.70.0 - (July 31, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 1.69.0 - (July 25, 2023)
 
 * Changed