+++ packages/terra-tabs/CHANGELOG.md
@@ -5,6 +5,11 @@
 * Added
  * Added `isDraggable` and `onTabOrderChange` props to support drag and drop feature in tabs.
 
+## 7.1.0 - (July 31, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 7.0.0 - (July 25, 2023)
 
 * Breaking Change